added more examples
